# Add prefetching (batched MultiGet) for experimental Ribbon filter