--- a/util/ribbon_alg.h
+++ b/util/ribbon_alg.h
@@ -1050,13 +1050,13 @@ void InterleavedBackSubst(InterleavedSolutionStorage *iss,
   std::unique_ptr<CoeffRow[]> state{new CoeffRow[num_columns]()};
 
   Index block = num_blocks;
-  Index segment = num_segments;
+  Index segment_num = num_segments;
   while (block > upper_start_block) {
     --block;
     BackSubstBlock(state.get(), num_columns, bs, block * kCoeffBits);
-    segment -= num_columns;
+    segment_num -= num_columns;
     for (Index i = 0; i < num_columns; ++i) {
-      iss->StoreSegment(segment + i, state[i]);
+      iss->StoreSegment(segment_num + i, state[i]);
     }
   }
   // Now (if applicable), region using lower number of columns
@@ -1066,23 +1066,23 @@ void InterleavedBackSubst(InterleavedSolutionStorage *iss,
   while (block > 0) {
     --block;
     BackSubstBlock(state.get(), num_columns, bs, block * kCoeffBits);
-    segment -= num_columns;
+    segment_num -= num_columns;
     for (Index i = 0; i < num_columns; ++i) {
-      iss->StoreSegment(segment + i, state[i]);
+      iss->StoreSegment(segment_num + i, state[i]);
     }
   }
   // Verify everything processed
   assert(block == 0);
-  assert(segment == 0);
+  assert(segment_num == 0);
 }
 
 // Prefetch memory for a key in InterleavedSolutionStorage.
 template <typename InterleavedSolutionStorage, typename PhsfQueryHasher>
-void InterleavedPrepareQuery(
+inline void InterleavedPrepareQuery(
     const typename PhsfQueryHasher::Key &key, const PhsfQueryHasher &hasher,
     const InterleavedSolutionStorage &iss,
     typename PhsfQueryHasher::Hash *saved_hash,
-    typename InterleavedSolutionStorage::Index *saved_segment,
+    typename InterleavedSolutionStorage::Index *saved_segment_num,
     typename InterleavedSolutionStorage::Index *saved_num_columns,
     typename InterleavedSolutionStorage::Index *saved_start_bit) {
   using Hash = typename PhsfQueryHasher::Hash;
@@ -1100,8 +1100,8 @@ void InterleavedPrepareQuery(
   const Index upper_start_block = iss.GetUpperStartBlock();
   Index num_columns = iss.GetUpperNumColumns();
   Index start_block_num = start_slot / kCoeffBits;
-  Index segment = start_block_num * num_columns -
-                  std::min(start_block_num, upper_start_block);
+  Index segment_num = start_block_num * num_columns -
+                      std::min(start_block_num, upper_start_block);
   // Change to lower num columns if applicable.
   // (This should not compile to a conditional branch.)
   num_columns -= (start_block_num < upper_start_block) ? 1 : 0;
@@ -1110,10 +1110,10 @@ void InterleavedPrepareQuery(
 
   Index segment_count = num_columns + (start_bit == 0 ? 0 : num_columns);
 
-  iss.PrefetchSegmentRange(segment, segment + segment_count);
+  iss.PrefetchSegmentRange(segment_num, segment_num + segment_count);
 
   *saved_hash = hash;
-  *saved_segment = segment;
+  *saved_segment_num = segment_num;
   *saved_num_columns = num_columns;
   *saved_start_bit = start_bit;
 }
@@ -1121,9 +1121,9 @@ void InterleavedPrepareQuery(
 // General PHSF query from InterleavedSolutionStorage, using data for
 // the query key from InterleavedPrepareQuery
 template <typename InterleavedSolutionStorage, typename PhsfQueryHasher>
-typename InterleavedSolutionStorage::ResultRow InterleavedPhsfQuery(
+inline typename InterleavedSolutionStorage::ResultRow InterleavedPhsfQuery(
     typename PhsfQueryHasher::Hash hash,
-    typename InterleavedSolutionStorage::Index segment,
+    typename InterleavedSolutionStorage::Index segment_num,
     typename InterleavedSolutionStorage::Index num_columns,
     typename InterleavedSolutionStorage::Index start_bit,
     const PhsfQueryHasher &hasher, const InterleavedSolutionStorage &iss) {
@@ -1143,14 +1143,14 @@ typename InterleavedSolutionStorage::ResultRow InterleavedPhsfQuery(
   ResultRow sr = 0;
   const CoeffRow cr_left = cr << start_bit;
   for (Index i = 0; i < num_columns; ++i) {
-    sr ^= BitParity(iss.LoadSegment(segment + i) & cr_left) << i;
+    sr ^= BitParity(iss.LoadSegment(segment_num + i) & cr_left) << i;
   }
 
   if (start_bit > 0) {
-    segment += num_columns;
+    segment_num += num_columns;
     const CoeffRow cr_right = cr >> (kCoeffBits - start_bit);
     for (Index i = 0; i < num_columns; ++i) {
-      sr ^= BitParity(iss.LoadSegment(segment + i) & cr_right) << i;
+      sr ^= BitParity(iss.LoadSegment(segment_num + i) & cr_right) << i;
     }
   }
 
@@ -1159,9 +1159,9 @@ typename InterleavedSolutionStorage::ResultRow InterleavedPhsfQuery(
 
 // Filter query a key from InterleavedFilterQuery.
 template <typename InterleavedSolutionStorage, typename FilterQueryHasher>
-bool InterleavedFilterQuery(
+inline bool InterleavedFilterQuery(
     typename FilterQueryHasher::Hash hash,
-    typename InterleavedSolutionStorage::Index segment,
+    typename InterleavedSolutionStorage::Index segment_num,
     typename InterleavedSolutionStorage::Index num_columns,
     typename InterleavedSolutionStorage::Index start_bit,
     const FilterQueryHasher &hasher, const InterleavedSolutionStorage &iss) {
@@ -1187,7 +1187,7 @@ bool InterleavedFilterQuery(
   // * get rid of start_bit == 0 condition with careful fetching & shifting
   if (start_bit == 0) {
     for (Index i = 0; i < num_columns; ++i) {
-      if (BitParity(iss.LoadSegment(segment + i) & cr) !=
+      if (BitParity(iss.LoadSegment(segment_num + i) & cr) !=
           (static_cast<int>(expected >> i) & 1)) {
         return false;
       }
@@ -1198,8 +1198,8 @@ bool InterleavedFilterQuery(
 
     for (Index i = 0; i < num_columns; ++i) {
       CoeffRow soln_data =
-          (iss.LoadSegment(segment + i) & cr_left) ^
-          (iss.LoadSegment(segment + num_columns + i) & cr_right);
+          (iss.LoadSegment(segment_num + i) & cr_left) ^
+          (iss.LoadSegment(segment_num + num_columns + i) & cr_right);
       if (BitParity(soln_data) != (static_cast<int>(expected >> i) & 1)) {
         return false;
       }